requirements.txt
# EconPaper Pro 依赖

# 核心框架
gradio>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# OpenAI兼容接口
openai>=1.30.0

# 向量数据库
chromadb>=0.4.0

# 文档解析
PyMuPDF>=1.23.0
python-docx>=0.8.11

# 学术搜索
scholarly>=1.7.0
serpapi>=0.1.0

# 文本处理
jieba>=0.42.1
# difflib is part of Python standard library, no need to install

# 工具库
httpx>=0.25.0
tenacity>=8.2.0
loguru>=0.7.0

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# 打包工具（开发时可选）
# pyinstaller>=6.0.0
